refactor: move methods from `UserService` to `AuthenticationService` and rename the current `AuthenticationService` to `AuthenticationManager`
