+++ src/main/kotlin/se/uulm/snowballr/backend/service/UserService.kt
@@ -42,7 +42,6 @@ import snowballr.UserSettingsOuterClass.UserSettings as GrpcUserSettings
 
 val Logger = KotlinLogging.logger { }
 
-@Suppress("ComplexInterface", "TooManyFunctions")
 interface IUserService {
     /**
      * Service implementation of [SnowballRService.getUserById].
@@ -68,21 +67,6 @@ interface IUserService {
      * Service implementation of [SnowballRService.register].
      */
     suspend fun register(request: Authentication.RegisterRequest): Base.Nothing
-
-    /**
-     * Service implementation of [SnowballRService.verifyEmail].
-     */
-    suspend fun verifyEmail(request: Authentication.VerifyEmailRequest): Base.Nothing
-
-    /**
-     * Service implementation of [SnowballRService.logout].
-     */
-    suspend fun logout(): Base.Nothing
-
-    /**
-     * Service implementation of [SnowballRService.login].
-     */
-    suspend fun login(request: Authentication.LoginRequest): Base.Nothing
 
     /**
      * Service implementation of [SnowballRService.updateUser].
@@ -118,16 +102,13 @@ private const val VERIFICATION_TOKEN_LENGTH = 48
  * @param projectMemberRepo The repository responsible for managing persistence operations for project members.
  * @param criterionRepo The repository responsible for managing persistence operations for criteria.
  * @param verificationTokenRepo The repository responsible for managing persistence operations for verification tokens.
- * @param jwtService The utility for handling JWT operations, such as token parsing and validation.
  * @param emailManager The manager responsible for sending emails.
  */
-@Suppress("TooManyFunctions")
 class UserService(
     private val userRepo: IUserTableRepo,
     private val projectMemberRepo: IProjectMemberTableRepo,
     private val criterionRepo: ICriterionTableRepo,
     private val verificationTokenRepo: IVerificationTokenTableRepo,
-    private val jwtService: IJwtService,
     private val emailManager: IEmailManager,
 ) : IUserService {
     companion object {
@@ -251,71 +232,6 @@ class UserService(
                 verificationLink,
             ),
         )
-
-        return Base.Nothing.getDefaultInstance()
-    }
-
-    override suspend fun verifyEmail(request: Authentication.VerifyEmailRequest): Base.Nothing {
-        val verificationToken = verificationTokenRepo.getVerificationTokenByValue(request.token)
-            ?: throw VerificationTokenNotFoundException()
-
-        // Check if the token has expired
-        if (OffsetDateTime.now().isAfter(verificationToken.expiresAt)) {
-            verificationTokenRepo.deleteVerificationToken(request.token)
-            throw VerificationTokenNotFoundException()
-        }
-
-        // Check whether the user exists
-        val user = userRepo.getUserById(verificationToken.userId).getOrThrow()
-
-        // Update the user's status to active
-        val updatedUser = user.copy(status = UserStatus.USER_STATUS_ACTIVE)
-        val userUpdate = GrpcUser.Update.newBuilder()
-            .setUser(updatedUser.toGrpcUser())
-            .setMask(FieldMask.newBuilder().addPaths("user.status").build())
-            .build()
-        userRepo.updateUser(userUpdate)
-
-        // Remove the verification token after successful verification
-        verificationTokenRepo.deleteVerificationToken(request.token)
-
-        return Base.Nothing.getDefaultInstance()
-    }
-
-    override suspend fun logout(): Base.Nothing {
-        GrpcContext.setAuthCookiesInContext("", "")
-
-        return Base.Nothing.getDefaultInstance()
-    }
-
-    override suspend fun login(request: Authentication.LoginRequest): Base.Nothing {
-        // Check whether a user with the given email exists
-        val user =
-            try {
-                userRepo.getUserByEmail(request.email).getOrThrow()
-            } catch (_: NotFoundException) {
-                throw UnauthenticatedException()
-            }
-
-        // Check whether the user is active (verified email)
-        if (user.status != UserStatus.USER_STATUS_ACTIVE) {
-            throw UnauthenticatedException()
-        }
-
-        // Verify the password against the stored hash
-        val storedPasswordHash = try {
-            userRepo.getPasswordHashByEmail(request.email).getOrThrow()
-        } catch (_: NotFoundException) {
-            throw UnauthenticatedException()
-        }
-
-        if (!PasswordUtils.verifyPassword(request.password, storedPasswordHash)) {
-            throw UnauthenticatedException()
-        }
-
-        // Generate JWT tokens
-        val (accessToken, refreshToken) = jwtService.generateAuthTokens(user.id)
-        GrpcContext.setAuthCookiesInContext(accessToken, refreshToken)
 
         return Base.Nothing.getDefaultInstance()
     }
