+++ src/test/kotlin/se/uulm/snowballr/backend/auth/AuthenticateTest.kt
@@ -1,4 +1,4 @@
-package se.uulm.snowballr.backend.service.authentication
+package se.uulm.snowballr.backend.auth
 
 import io.jsonwebtoken.JwtException
 import io.mockk.every
@@ -6,10 +6,7 @@ import io.mockk.mockk
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import se.uulm.snowballr.backend.GrpcTestContextExtension
-import se.uulm.snowballr.backend.auth.GrpcContext
-import se.uulm.snowballr.backend.auth.JwtService
 import se.uulm.snowballr.backend.model.jwt.ParsedJwtAuthClaims
-import se.uulm.snowballr.backend.service.AuthenticationService
 import snowballr.Authentication
 import java.util.Date
 import java.util.UUID
@@ -19,11 +16,11 @@ import kotlin.test.assertTrue
 
 @ExtendWith(GrpcTestContextExtension::class)
 class AuthenticateTest {
-    private val jwtServiceMock = mockk<JwtService> {
-        every { getAccessTokenTTL() } returns JwtService.ACCESS_TOKEN_EXPIRATION_MS
-        every { getRefreshTokenTTL() } returns JwtService.REFRESH_TOKEN_EXPIRATION_MS
+    private val jwtServiceMock = mockk<JwtManager> {
+        every { getAccessTokenTTL() } returns JwtManager.ACCESS_TOKEN_EXPIRATION_MS
+        every { getRefreshTokenTTL() } returns JwtManager.REFRESH_TOKEN_EXPIRATION_MS
     }
-    private val authenticationService = AuthenticationService(jwtServiceMock)
+    private val authenticationService = AuthenticationManager(jwtServiceMock)
 
     @Test
     fun `When access token is valid, then authentication succeeds`() {
