+++ src/test/kotlin/se/uulm/snowballr/backend/service/authentication/LoginTest.kt
@@ -1,4 +1,4 @@
-package se.uulm.snowballr.backend.service.user
+package se.uulm.snowballr.backend.service.authentication
 
 import io.mockk.coEvery
 import io.mockk.every
@@ -37,10 +37,10 @@ class LoginTest : MainServiceTest() {
     fun `When a user with an unconfirmed email tries to log in, then an UnauthenticatedException is thrown`() =
         runTest {
             val testUser = DataBuilder.createExampleUser(status = UserStatus.USER_STATUS_ACTIVE_UNCONFIRMED)
-            val request = LoginRequest.newBuilder().apply {
-                email = testUser.email
-                password = "anyPassword"
-            }.build()
+            val request = LoginRequest.newBuilder()
+                .setEmail(testUser.email)
+                .setPassword("anyPassword")
+            .build()
 
             coEvery { userRepoMock.getUserByEmail(testUser.email) } returns Result.success(testUser)
 
@@ -50,10 +50,10 @@ class LoginTest : MainServiceTest() {
     @Test
     fun `When a deleted user tries to log in, then an UnauthenticatedException is thrown`() = runTest {
         val testUser = DataBuilder.createExampleUser(status = UserStatus.USER_STATUS_DELETED)
-        val request = LoginRequest.newBuilder().apply {
-            email = testUser.email
-            password = "anyPassword"
-        }.build()
+        val request = LoginRequest.newBuilder()
+            .setEmail(testUser.email)
+            .setPassword("anyPassword")
+            .build()
 
         coEvery { userRepoMock.getUserByEmail(testUser.email) } returns Result.success(testUser)
 
@@ -64,10 +64,10 @@ class LoginTest : MainServiceTest() {
     fun `When a user with an unspecified status tries to log in, then an UnauthenticatedException is thrown`() =
         runTest {
             val testUser = DataBuilder.createExampleUser(status = UserStatus.USER_STATUS_UNSPECIFIED)
-            val request = LoginRequest.newBuilder().apply {
-                email = testUser.email
-                password = "anyPassword"
-            }.build()
+            val request = LoginRequest.newBuilder()
+                .setEmail(testUser.email)
+                .setPassword("anyPassword")
+                .build()
 
             coEvery { userRepoMock.getUserByEmail(testUser.email) } returns Result.success(testUser)
 
@@ -78,10 +78,10 @@ class LoginTest : MainServiceTest() {
     fun `When the password hash cannot be retrieved, then an UnauthenticatedException is thrown`() = runTest {
         val testUser = DataBuilder.createExampleUser(status = UserStatus.USER_STATUS_ACTIVE)
 
-        val request = LoginRequest.newBuilder().apply {
-            email = testUser.email
-            password = "anyPassword"
-        }.build()
+        val request = LoginRequest.newBuilder()
+            .setEmail(testUser.email)
+            .setPassword("anyPassword")
+            .build()
 
         coEvery { userRepoMock.getUserByEmail(testUser.email) } returns Result.success(testUser)
         val exception = NotFoundException(EntityType.USER, testUser.email, identifierType = IdentifierType.EMAIL)
@@ -96,10 +96,10 @@ class LoginTest : MainServiceTest() {
         val userPassword = "AAbb__00"
         val passwordHash = PasswordUtils.hashPassword(userPassword)
 
-        val request = LoginRequest.newBuilder().apply {
-            email = testUser.email
-            password = "wrongPassword"
-        }.build()
+        val request = LoginRequest.newBuilder()
+            .setEmail(testUser.email)
+            .setPassword("wrongPassword")
+            .build()
 
         coEvery { userRepoMock.getUserByEmail(testUser.email) } returns Result.success(testUser)
         coEvery { userRepoMock.getPasswordHashByEmail(testUser.email) } returns Result.success(passwordHash)
@@ -114,10 +114,10 @@ class LoginTest : MainServiceTest() {
         val passwordHash = PasswordUtils.hashPassword(userPassword)
         val tokens = JwtAuthTokens("accessToken", "refreshToken")
 
-        val request = LoginRequest.newBuilder().apply {
-            email = testUser.email
-            password = userPassword
-        }.build()
+        val request = LoginRequest.newBuilder()
+            .setEmail(testUser.email)
+            .setPassword(userPassword)
+            .build()
 
         coEvery { userRepoMock.getUserByEmail(testUser.email) } returns Result.success(testUser)
         coEvery { userRepoMock.getPasswordHashByEmail(testUser.email) } returns Result.success(passwordHash)
